% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hmm_movement.R
\name{d_wrapped_normal}
\alias{d_wrapped_normal}
\title{Wrapped-normal density}
\usage{
d_wrapped_normal(theta, mean, sd, log = FALSE, n_terms = 5L)
}
\arguments{
\item{theta}{Numeric vector of angles.}

\item{mean}{Circular mean.}

\item{sd}{Standard deviation before wrapping.}

\item{log}{Logical. Return log-density if \code{TRUE}.}

\item{n_terms}{Number of wrapped normal series terms on each side of zero.}
}
\value{
A numeric vector of densities or log-densities.
}
\description{
Wrapped-normal density
}
\examples{
d_wrapped_normal(0, mean = 0, sd = 1)
}
