% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{package}
\name{evalueHMM-package}
\alias{evalueHMM}
\alias{evalueHMM-package}
\title{evalueHMM: predictive e-diagnostics for movement HMMs}
\description{
The package provides tools for constructing predictive e-process diagnostics
for hidden Markov models used with animal movement data. Core functions cover
simulation, observable predictive densities by filtering, e-process
construction, diagnostic alternatives, localization, combinations and plotting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AurelienNicosiaULaval/predictive_e_diagnostics_hmm}
  \item Report bugs at \url{https://github.com/AurelienNicosiaULaval/predictive_e_diagnostics_hmm/issues}
}

}
\author{
\strong{Maintainer}: Aurelien Nicosia \email{aurelien.nicosia@ulaval.ca}

}
\keyword{internal}
