% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_copula.R
\name{gaussian_copula_log_density}
\alias{gaussian_copula_log_density}
\title{Gaussian copula log-density}
\usage{
gaussian_copula_log_density(u, v, rho)
}
\arguments{
\item{u, v}{Numeric vectors with values in \verb{[0, 1]}.}

\item{rho}{Gaussian copula correlation parameter in \verb{(-1, 1)}.}
}
\value{
A numeric vector of log-density values.
}
\description{
Gaussian copula log-density
}
