% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hmm_movement.R
\name{p_wrapped_normal}
\alias{p_wrapped_normal}
\title{Wrapped-normal distribution function}
\usage{
p_wrapped_normal(theta, mean, sd, n_terms = 8L)
}
\arguments{
\item{theta}{Numeric vector of angles.}

\item{mean}{Circular mean.}

\item{sd}{Standard deviation before wrapping.}

\item{n_terms}{Number of wrapped normal series terms on each side of zero.}
}
\value{
A numeric vector of distribution function values clipped to \verb{[0, 1]}.
}
\description{
Wrapped-normal distribution function
}
\examples{
p_wrapped_normal(0, mean = 0, sd = 1)
}
