% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_eprocess}
\alias{plot_eprocess}
\title{Plot a predictive e-process}
\usage{
plot_eprocess(eprocess, title = "Predictive e-process")
}
\arguments{
\item{eprocess}{Object returned by \code{\link[=compute_eprocess]{compute_eprocess()}}.}

\item{title}{Plot title.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a predictive e-process
}
\examples{
eprocess <- compute_eprocess(c(-1, -1), c(-0.8, -0.9))
plot_eprocess(eprocess)
}
