% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_long_horizon.R
\name{simulate_block_straightness_values}
\alias{simulate_block_straightness_values}
\title{Simulate block straightness values}
\usage{
simulate_block_straightness_values(
  n_simulations,
  block_size,
  parameters,
  generator = c("hmm", "ar_angle"),
  rho_by_state = NULL
)
}
\arguments{
\item{n_simulations}{Number of simulated blocks.}

\item{block_size}{Number of observations per block.}

\item{parameters}{Parameter list from \code{\link[=create_hmm_movement_parameters]{create_hmm_movement_parameters()}}.}

\item{generator}{Either \code{"hmm"} or \code{"ar_angle"}.}

\item{rho_by_state}{State-dependent autoregressive correlations for
\code{generator = "ar_angle"}.}
}
\value{
A numeric vector of straightness values.
}
\description{
Simulate block straightness values
}
