% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hmm_movement.R
\name{simulate_wrapped_normal}
\alias{simulate_wrapped_normal}
\title{Simulate wrapped-normal angles}
\usage{
simulate_wrapped_normal(n, mean, sd)
}
\arguments{
\item{n}{Number of angles to simulate.}

\item{mean}{Circular mean.}

\item{sd}{Standard deviation before wrapping.}
}
\value{
A numeric vector of angles in \eqn{[-\pi, \pi)}.
}
\description{
Simulate wrapped-normal angles
}
