% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hmm_training.R
\name{summarise_hmm_parameter_error}
\alias{summarise_hmm_parameter_error}
\title{Summarise parameter error for simulated HMMs}
\usage{
summarise_hmm_parameter_error(estimated_parameters, true_parameters)
}
\arguments{
\item{estimated_parameters}{Estimated parameter list.}

\item{true_parameters}{True parameter list.}
}
\value{
A one-row data frame of error summaries.
}
\description{
Summarise parameter error for simulated HMMs
}
