% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_interface.R
\name{summarise_predictive_diagnostic}
\alias{summarise_predictive_diagnostic}
\title{Summarise predictive e-diagnostic results.}
\usage{
summarise_predictive_diagnostic(diagnostic)
}
\arguments{
\item{diagnostic}{A \code{predictive_e_diagnostic} object.}
}
\value{
A data frame summarizing the diagnostic.
}
\description{
Summarise predictive e-diagnostic results.
}
\examples{
diagnostic <- make_predictive_diagnostic(
  diagnostic_name = "toy",
  log_p0 = c(-1.2, -0.8),
  log_q = c(-1.0, -0.7)
)
summarise_predictive_diagnostic(diagnostic)
}
