% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_interface.R
\name{print.predictive_e_diagnostic}
\alias{print.predictive_e_diagnostic}
\title{Print method for predictive_e_diagnostic.}
\usage{
\method{print}{predictive_e_diagnostic}(x, ...)
}
\arguments{
\item{x}{A \code{predictive_e_diagnostic} object.}

\item{...}{Unused.}
}
\description{
Print method for predictive_e_diagnostic.
}
\keyword{internal}
